# [F#] Native compilation and speedup for JIT and native