--- a/fsharp/linux-native.sh
+++ b/fsharp/linux-native.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+RID=linux-x64
+
+dotnet publish -r $RID -c Release trace.fsproj &&
+bin/Release/netcoreapp2.1/$RID/publish/trace $@