Manually inlining `scattered` really helped in this case, a lot of memory traffic from the tuples and optionals went away. Moving some functions from inline back to non inlined helps code quality because it allows the JIT (inline is forced by the compiler) to make a decision whether to inline or not based on method/function body size, some were already pretty chunky. Similarly moving `objsHit` back to `Option` from `ValueOption` helps because the unboxed values to pass along the recursive calls are otherwise quite big, causing stack spills and memory traffic.

Console.WriteLine is used because printfn requires reflection which complicates the native build.
netcoreapp2.1 is used because frankly it performs better in both native and jitted cases (by about 10 to 15%) . I'll investigate and ask around.

After: 
```
$ dotnet run -c release --  -f rgbbox_1000.ppm -s rgbbox -n 1000 -m 1000
Using scene 'rgbbox' (-s to switch).
Using 1 warmup runs before benchmarking (-r to change).
Scene BVH construction in 0.0003179s.
Rendering in 1.1639827s.
Writing image to rgbbox_1000.ppm.
```

Before (also one warmup run): 
```
$ dotnet run -c release --  -f rgbbox_1000.ppm -s rgbbox -n 1000 -m 1000 -r 1                                               
Using scene 'rgbbox' (-s to switch).
Using 1 warmup runs before benchmarking (-r to change).
Scene BVH construction in 0.000546s.
Rendering in 1.688101s.
Writing image to rgbbox_1000.ppm.
```

Native is even faster.