@athas 2020-04-15
I'm sorry the line is so blurry, but inlining `scatter` is beyond my tolerance for hand-optimisation.  That would be hard to maintain in a more general case where we have more kinds of materials.  The other F# improvements have purely been about adding annotations or using value types instead of reference types, none of which have changed the feel, structure, or readability of the code.  I think that shows F# in a much better light.

I'm perfectly happy with the native compilation and other changes, so I'll merge the PR if you restrict it to that.